git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/docvalues@1029003 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/search/cache/DocTermsCreator.java
@@ -164,7 +164,7 @@ public class DocTermsCreator extends EntryCreatorWithOptions<DocTerms>
 
     @Override
     public BytesRef getTerm(int docID, BytesRef ret) {
-      final int pointer = (int) docToOffset.get(docID);
+      final long pointer = docToOffset.get(docID);
       return bytes.fillUsingLengthPrefix(ret, pointer);
     }
   }
